improving perfomance
