+++ lib/Login.dart
@@ -1,8 +1,8 @@
-import 'package:ar_pin/Cadastro.dart';
+import 'package:ar_pin/cadastro_view.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'HomePage.dart';
+import 'home_page.dart';
 class Login extends StatefulWidget {
   const Login({super.key});
 
@@ -22,8 +22,10 @@ backgroundColor: Colors.white,
           double baseWidth = 1040;
           double fem = constraints.maxWidth / baseWidth;
           double ffem = fem * 0.97;
+          
 
-          return Container(
+          //container ou sized box?
+          return SizedBox(
             width: MediaQuery.of(context).size.width,
             height: MediaQuery.of(context).size.height,
             child: Container(
@@ -57,10 +59,10 @@ backgroundColor: Colors.white,
                             fontWeight: FontWeight.w300,
                             height: 1.5 * ffem / fem,
                             fontStyle: FontStyle.italic,
-                            color: Color(0xff000000),
+                            color: const Color(0xff000000),
                           ),
                           children: [
-                            TextSpan(
+                            const TextSpan(
                               text: 'AR-',
                             ),
                             TextSpan(
@@ -70,7 +72,7 @@ backgroundColor: Colors.white,
                                 fontWeight: FontWeight.w600,
                                 height: 1.5 * ffem / fem,
                                 fontStyle: FontStyle.italic,
-                                color: Color(0xff000000),
+                                color: const Color(0xff000000),
                               ),
                             ),
                           ],
@@ -82,7 +84,7 @@ backgroundColor: Colors.white,
                       width: 919 * fem,
                       height: 125 * fem,
                       decoration: BoxDecoration(
-                        color: Color(0x7fd9d1e4),
+                        color: const Color(0x7fd9d1e4),
                         borderRadius: BorderRadius.circular(50 * fem),
                       ),
                       child: TextField(
@@ -93,13 +95,13 @@ backgroundColor: Colors.white,
                           enabledBorder: InputBorder.none,
                           contentPadding: EdgeInsets.fromLTRB(25 * fem, 0 * fem, 0 * fem, 0 * fem),
                           hintText: 'Email',
-                          hintStyle: TextStyle(color: Color(0xff000000)),
+                          hintStyle: const TextStyle(color: Color(0xff000000)),
                         ),
                         style: GoogleFonts.poppins(
                           fontSize: 48 * ffem,
                           fontWeight: FontWeight.w400,
                           height: 1.5 * ffem / fem,
-                          color: Color(0xff000000),
+                          color: const Color(0xff000000),
                         ),
                       ),
                     ),
@@ -108,7 +110,7 @@ backgroundColor: Colors.white,
                       width: 919 * fem,
                       height: 125 * fem,
                       decoration: BoxDecoration(
-                        color: Color(0x7fd9d1e4),
+                        color: const Color(0x7fd9d1e4),
                         borderRadius: BorderRadius.circular(50 * fem),
                       ),
                       child: TextField(
@@ -119,13 +121,13 @@ backgroundColor: Colors.white,
                           enabledBorder: InputBorder.none,
                           contentPadding: EdgeInsets.fromLTRB(25 * fem, 0 * fem, 0 * fem, 0 * fem),
                           hintText: 'Senha',
-                          hintStyle: TextStyle(color: Color(0xff000000)),
+                          hintStyle: const TextStyle(color: Color(0xff000000)),
                         ),
                         style: GoogleFonts.poppins(
                           fontSize: 48 * ffem,
                           fontWeight: FontWeight.w400,
                           height: 1.5 * ffem / fem,
-                          color: Color(0xff000000),
+                          color: const Color(0xff000000),
                         ),
                       ),
                     ),
@@ -133,7 +135,8 @@ backgroundColor: Colors.white,
                       margin: EdgeInsets.fromLTRB(0 * fem, 45 * fem, 0 * fem, 25 * fem),
                       child: ElevatedButton(
                         style: ElevatedButton.styleFrom(
-                          primary: Color(0xffe51f43),
+                          //primary ou background color?
+                          backgroundColor: const Color(0xffe51f43),
                           minimumSize: Size(919 * fem, 125 * fem),
                           elevation: 0,
                           shape: const RoundedRectangleBorder(borderRadius: BorderRadius.all(Radius.circular(50))),
@@ -141,7 +144,7 @@ backgroundColor: Colors.white,
                         onPressed: () {
                           FirebaseAuth.instance.signInWithEmailAndPassword(email: _emailTextController.text, password: _passwordTextController.text).then((value){
                             Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => const HomePage())).onError((error, stackTrace) {
-                              print(error);
+                              // print(error);
                             });
                           });
                         },
@@ -160,8 +163,8 @@ backgroundColor: Colors.white,
                       margin: EdgeInsets.fromLTRB(0 * fem, 0 * fem, 0 * fem, 50 * fem),
                       child: OutlinedButton(
                         style: OutlinedButton.styleFrom(
-                          primary: Color(0xffe51f43),
-                          side: BorderSide(color: Color(0xffe51f43), width: 1),
+                          //foreground ou primary?
+                          foregroundColor: const Color(0xffe51f43), side: const BorderSide(color: Color(0xffe51f43), width: 1),
                           minimumSize: Size(919 * fem, 125 * fem),
                           shape: const RoundedRectangleBorder(borderRadius: BorderRadius.all(Radius.circular(50))),
                         ),
@@ -174,7 +177,7 @@ backgroundColor: Colors.white,
                             fontSize: 48 * ffem,
                             fontWeight: FontWeight.w400,
                             height: 1.5 * ffem / fem,
-                            color: Color(0xffe51f43),
+                            color: const Color(0xffe51f43),
                           ),
                         ),
                       ),
@@ -183,7 +186,7 @@ backgroundColor: Colors.white,
                       margin: EdgeInsets.fromLTRB(0 * fem, 200 * fem, 0 * fem, 0 * fem),
                       child: TextButton(
                         onPressed: () {
-                          Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (context) => Cadastro()));
+                          Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (context) => const Cadastro()));
                         },
                         style: TextButton.styleFrom(
                           padding: EdgeInsets.zero,
@@ -194,8 +197,8 @@ backgroundColor: Colors.white,
                             fontSize: 45 * ffem,
                             fontWeight: FontWeight.w600,
                             height: 1.5 * ffem / fem,
-                            color: Color(0xff2473ea),
-                            decorationColor: Color(0xff2473ea),
+                            color: const Color(0xff2473ea),
+                            decorationColor: const Color(0xff2473ea),
                           ),
                         ),
                       ),
