+++ lib/baixar_modelo.dart
@@ -1,9 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'HomePage.dart';
-import 'UserProfile.dart';
+import 'home_page.dart';
+import 'user_profile.dart';
 
 class BXModelo extends StatelessWidget {
+  //testar se influencia ou não a aplicação
+  const BXModelo({super.key});
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -37,13 +40,14 @@ class BXModelo extends StatelessWidget {
           double fem = constraints.maxWidth / baseWidth;
           double ffem = fem * 0.97;
 
-          return Container(
+          //voltar para o layout original(container?)
+          return SizedBox(
             width: double.infinity,
             child: Container(
               // visualizarprojetocqD (255:42)
               padding: EdgeInsets.fromLTRB(0 * fem, 50 * fem, 0 * fem, 0 * fem),
               width: double.infinity,
-              decoration: BoxDecoration(
+              decoration: const BoxDecoration(
                 color: Color(0xffffffff),
               ),
               child: Column(
@@ -59,7 +63,7 @@ class BXModelo extends StatelessWidget {
                           fontWeight: FontWeight.w700,
                           height: 1.5 * ffem / fem,
                           fontStyle: FontStyle.italic,
-                          color: Color(0xff000000),
+                          color: const Color(0xff000000),
                         ),
                       ),
                     ),
@@ -72,7 +76,7 @@ class BXModelo extends StatelessWidget {
                           fontSize: 52 * ffem,
                           fontWeight: FontWeight.w400,
                           height: 1.5 * ffem / fem,
-                          color: Color(0xff000000),
+                          color: const Color(0xff000000),
                         ),
                       ),
                     ),
@@ -89,7 +93,7 @@ class BXModelo extends StatelessWidget {
                           margin: EdgeInsets.fromLTRB(0 * fem, 0 * fem, 25 * fem, 250 * fem),
                           width: 644 * fem,
                           height: 596 * fem,
-                          decoration: BoxDecoration(
+                          decoration: const BoxDecoration(
                             image: DecorationImage(
                               fit: BoxFit.cover,
                               image: AssetImage(
@@ -104,7 +108,7 @@ class BXModelo extends StatelessWidget {
                           padding: EdgeInsets.fromLTRB(34.5 * fem, 27 * fem, 46 * fem, 27 * fem),
                           width: double.infinity,
                           decoration: BoxDecoration(
-                            color: Color(0xffe51f43),
+                            color: const Color(0xffe51f43),
                             borderRadius: BorderRadius.circular(50 * fem),
                           ),
                           child: Row(
@@ -120,11 +124,12 @@ class BXModelo extends StatelessWidget {
                                     fontSize: 40 * ffem,
                                     fontWeight: FontWeight.w600,
                                     height: 1.5 * ffem / fem,
-                                    color: Color(0xffffffff),
+                                    color: const Color(0xffffffff),
                                   ),
                                 ),
                               ),
-                              Container(
+                              //container ou sizedbox?
+                              SizedBox(
                                 // download11MkB (440:3)
                                 width: 60 * fem,
                                 height: 60 * fem,
