+++ lib/quest_alter.dart
@@ -1,17 +1,17 @@
-import 'package:ar_pin/QuestAlter.dart';
+import 'package:ar_pin/quest_add.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'HomePage.dart';
-import 'UserProfile.dart';
+import 'home_page.dart';
+import 'user_profile.dart';
 
-class QuestAdd extends StatefulWidget {
-  const QuestAdd({super.key});
+class QuestAlter extends StatefulWidget {
+  const QuestAlter({super.key});
 
   @override
-  State<QuestAdd> createState() => _QuestAddState();
+  State<QuestAlter> createState() => _QuestAlterState();
 }
 
-class _QuestAddState extends State<QuestAdd> {
+class _QuestAlterState extends State<QuestAlter> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -50,7 +50,7 @@ class _QuestAddState extends State<QuestAdd> {
               Container(
                 margin: EdgeInsets.fromLTRB(0 * fem, 50 * fem, 0 * fem, 25 * fem),
                 child: Text(
-                  'Questões',
+                  'Questão 1',
                   textAlign: TextAlign.center,
                   style: GoogleFonts.poppins(
                     fontSize: 70 * ffem,
@@ -62,6 +62,7 @@ class _QuestAddState extends State<QuestAdd> {
                 ),
               ),
               _question(fem, ffem),
+              _answer(fem, ffem),
               _button(fem, ffem),
             ],
           );
@@ -72,8 +73,68 @@ class _QuestAddState extends State<QuestAdd> {
 
   _question(double fem, double ffem) {
     return Column(
-      crossAxisAlignment: CrossAxisAlignment.center,
       children: [
+        Container(
+          margin: EdgeInsets.fromLTRB(50 * fem, 0 * fem, 0 * fem, 0 * fem),
+          child: Align(
+            alignment: Alignment.centerLeft,
+            child: Text(
+              'Pergunta:',
+              textAlign: TextAlign.justify,
+              style: GoogleFonts.poppins(
+                fontSize: 64 * ffem,
+                fontWeight: FontWeight.w600,
+                height: 1.5 * ffem / fem,
+                fontStyle: FontStyle.italic,
+                color: const Color(0xff000000),
+              ),
+            ),
+          ),
+        ),
+        Container(
+          margin: EdgeInsets.fromLTRB(0 * fem, 0 * fem, 0 * fem, 0 * fem),
+          constraints: BoxConstraints(
+            maxWidth: 850 * fem,
+          ),
+          child: Align(
+            alignment: Alignment.centerLeft,
+            child: Text(
+              'Qual é a função principal de um LED?',
+              textAlign: TextAlign.justify,
+              style: GoogleFonts.poppins(
+                fontSize: 58 * ffem,
+                fontWeight: FontWeight.w300,
+                height: 1.5 * ffem / fem,
+                fontStyle: FontStyle.italic,
+                color: const Color(0xff000000),
+              ),
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+
+  _answer(double fem, double ffem) {
+    return Column(
+      children: [
+        Container(
+          margin: EdgeInsets.fromLTRB(50 * fem, 50 * fem, 0 * fem, 50 * fem),
+          child: Align(
+            alignment: Alignment.centerLeft,
+            child: Text(
+              'Respostas:',
+              textAlign: TextAlign.justify,
+              style: GoogleFonts.poppins(
+                fontSize: 64 * ffem,
+                fontWeight: FontWeight.w600,
+                height: 1.5 * ffem / fem,
+                fontStyle: FontStyle.italic,
+                color: const Color(0xff000000),
+              ),
+            ),
+          ),
+        ),
         Container(
           margin: EdgeInsets.fromLTRB(0 * fem, 0 * fem, 0 * fem, 35 * fem),
           width: 919 * fem,
@@ -81,32 +142,23 @@ class _QuestAddState extends State<QuestAdd> {
           child: ElevatedButton(
             style: ElevatedButton.styleFrom(
               primary: const Color(0x50DAD1E4),
-              minimumSize: Size(919 * fem, 125 * fem),
               elevation: 0,
               shape: const RoundedRectangleBorder(borderRadius: BorderRadius.all(Radius.circular(50))),
             ),
             onPressed: () {},
-            child: Row(
-              children: [
-                Expanded(
-                  child: Align(
-                    alignment: Alignment.centerLeft,
-                    child: Text(
-                      'Material Didatico',
-                      style: GoogleFonts.poppins(
-                        fontSize: 48 * ffem,
-                        fontWeight: FontWeight.w300,
-                        height: 1.5 * ffem / fem,
-                        color: Colors.black,
-                      ),
-                    ),
+            child: Expanded(
+              child: Align(
+                alignment: Alignment.centerLeft,
+                child: Text(
+                  'Gerar calor',
+                  style: GoogleFonts.poppins(
+                    fontSize: 58 * ffem,
+                    fontWeight: FontWeight.w300,
+                    height: 1.5 * ffem / fem,
+                    color: Colors.black,
                   ),
                 ),
-                Align(
-                  alignment: Alignment.centerRight,
-                  child: IconButton(iconSize: 0, onPressed: () {}, icon: Image.asset('assets/images/bin.png')),
-                ),
-              ],
+              ),
             ),
           ),
         ),
@@ -130,10 +182,10 @@ class _QuestAddState extends State<QuestAdd> {
                   shape: const RoundedRectangleBorder(borderRadius: BorderRadius.all(Radius.circular(50))),
                 ),
                 onPressed: () {
-                  Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => QuestAlter()));
+                  Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => QuestAdd()));
                 },
                 child: Text(
-                  'Adicionar Pergunta',
+                  'Salvar',
                   style: GoogleFonts.poppins(
                     fontSize: 48 * ffem,
                     fontWeight: FontWeight.w300,
