+++ lib/quest_conclude.dart
@@ -1,9 +1,9 @@
-import 'package:ar_pin/Tutorial.dart';
+import 'package:ar_pin/tutorial.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'HomePage.dart';
-import 'UserProfile.dart';
-import 'Quest.dart';
+import 'home_page.dart';
+import 'user_profile.dart';
+import 'quest.dart';
 
 class QuestConcluido extends StatelessWidget {
   final int score;
